Archetypes - Add test execution in IT (exec approach)
Change the goals of the integration tests of the archetypes to execute the
run.sh/.bat script with the task build_test to compile, run and execute the
integration tests of the generated projects.

To do this we need to make the run.sh file executable adding a post generate
groovy script to add the execution permission to the script file.